requirements.txt
flask
python-dotenv
sqlalchemy
pymysql
pandas
matplotlib
langchain
langchain-community
langchain-openai
